project/openclawd_stack/control/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
httpx
python-dotenv
gspread
google-auth
